+++ test/unit/fixtures/Basic.vue
@@ -1,0 +1,20 @@
+<template>
+  <div class="app__component">
+    <button>
+      Click
+
+      <MaterialRipple ref="ripple"></MaterialRipple>
+    </button>
+  </div>
+</template>
+
+<script>
+  import MaterialRipple from 'components/MaterialRipple'
+
+  export default {
+    name: 'app',
+    components: {
+      MaterialRipple
+    }
+  }
+</script>